This commit adds the dye tracer capabilities to POP.
I developed one category for tracers (surfdyes) that enter the ocean
along with precipitation, river runoff, sea ice melt, and glacial
runoff; and one category (sectdyes) for dyes that are defined at
internal ocean sections by continuously resetting their value to
1. These sections are specified in gx1v6_sectdyes_contents for the
gx1v6 grid, but separate files will need to be generated for other
grids (including the 0.3).

For these internal dye tracers, several sections can be assigned to
the same dye tracer, as specified in the sectdyes_contents file. This
allows for i) 're-using' tracers for different areas of interest (say,
the Arctic and Southern Ocean domains), but also ii) for piecing
together different subsections; for instance, the South Atlantic
section at 30S which straddles the 'seam' of the model domain; or
oblique sections, like the Barents Sea Opening (see the example in the
sectdyes_contents files).

These capabilities are activated by adding sectdyes and/or surfdyes in
the setting of OCN_TRACER_MODULES in env_build.xml. The number of
tracers (currently set to 6 for sectdyes, 4 for surfdyes) can be
changed in ACME-hilat/models/ocn/pop2/bld/pop2.buildexe.csh.