Update cice to the zbgc_colpkg.r1121
This version of cice has already been merged into RASM
and the MPAS-CICE branch.
It includes coupling hooks for ice to ocean plus bug fixes
and tuning of the bgc.  These are BFB with biogeochemistry off.
The nonBFB change is a bug fix in ice_shortwave.F90 that
has already been implemented and tested in the CESM and RASM.

This PR also turns off the default option of aerosols in sea ice
(both in the build and in the namelist).  This is a nonBFB change
as the aerosols alter the radiative properties of sea ice.