Added the IRF capability.
It also includes a change to the filtering procedure to make sure that
IRF values of 1 are not changed due to varthick. This leads to non-bfb
comparisons.

Here is Keith's explanation:

The filtering, usually referred to as time-averaging in POP, is done to
remove the computational mode that arises from using leap-frog
timestepping. What changed with the introduction of IRF was the
implementation of the filtering. Before IRF was brought in,
time-averaging code in the top layer, when using var-thick, was
effectively:

TRACER(old) = p5*(TRACER(old)*(dz(1)+SSH(old)) +
TRACER(cur)*(dz(1)+SSH(cur)))
TRACER(cur) = p5*(TRACER(cur)*(dz(1)+SSH(cur)) +
TRACER(new)*(dz(1)+SSH(new)))
SSH(old) = p5*(SSH(old)+SSH(cur))
SSH(cur) = p5*(SSH(cur)+SSH(new))
TRACER(old) = TRACER(old)/(dz(1)+SSH(old))
TRACER(cur) = TRACER(cur)/(dz(1)+SSH(cur))

This thickness-weighted time-averaging is done to ensure
TRACER*(dz(1)+SSH) is conserved when applying time-averaging.

The problem is that if TRACER(old), TRACER(cur), TRACER(new) are all
identical, say 1.0, the above code yields new values for TRACER(old) and
TRACER(cur) that are different from 1.0, because of roundoff error. In
infinite precision arithmetic, the filtered values would be identically
1.0. This situation of having identical tracer values at each time level
is atypical for most tracers, but it does arise for IRF tracers. It
seemed undesirable for time-averaging to lead to loss of constancy for
IRF tracers.

The new implementation adds limiters on thickness-weighted time-averaged
TRACER values to ensure that the filtered values do not exceed the min
or max of the pre-time-averaged values. The code was rearranged in order
to support these limiters.