Pseudotracers implementation
The changes in this commit enable the use of pseudotracers; in
particular pseudo-salinity and pseudo-temperature. These tracers
behave exactly the same as regular temperature and salinity, but can
eventually be perturbed with surface fluxes that regular temperature
and salinity won't see. So we can study the impact of surface flux
perturbations on the (pseudo-)stratification and vertical mixing,
without affecting the ocean circulation.

The package is enabled by specifying 'pseudotracers' in the
environmental variable OCN_TRACER_MODULES in env_build.xml. The
pseudo-temperature and pseudo-salinity generate their own vertical
mixing fields (output as VDC_pT and VDC_pS) and associated mixed layer
diagnostics (pHMXL, pHBLT, etc.).

A namelist switch 'use_pvdc_for_passive_tracers' is introduced that
allows the user to choose whether additional tracers (e.g. ecosys) are
subject to the regular diffusion or the pseudo-diffusion fields. This
also affects the KPP_SRC term. The selection is stored in the array
VDC_pick(1:nt), where a value of 1 for a certain tracer indicates that
that tracer is diffused by VDC_T, 2 by VDC_S, 3 by VDC_pT, and 4
by VDC_pS. Default behavior is (1, 2, 3, 4, 2,...) for (T, S, pT, pS,
other tracers...), so passive tracers are diffused by VDC_S. This is also
default behavior if no pseudotracers are selected.  If
use_pvdc_for_passive_tracers = true this behavior is changed to (1, 2,
3, 4, 4,...).