Locally-referenced virtual salt flux.
These changes allow for ROFF_F, IOFF_F and MELT_F to be converted to
virtual salt fluxes with reference to local salinity, instead of the
global reference salinity. This limits salt extraction in narrow
estuaries, hence preventing strongly negative salinities.

This capability is toggled with a new namelist variable:

 lroff_ref_to_local_salinity

Default is true for tx0.3v2, false for other grids.

There are two important notes:

1) The conversion from ROFF_F, IOFF_F and MELT_F to virtual salt flux
takes place just after POP receives the information from the coupler,
and hence depends on the local salinity /at that time/. It is updated as
often as the coupling interval, which is 1 day for the gx1v6 case, and 1
hour in the tx0.3v2 case. In contrast, the time-averaging of these
fluxes takes place at every time step, with the local salinities /at
that time/. The time-average of SFWF is hence slightly different than
the sum of the time-averages of the ROFF_F etc. This difference is a few
tenths of a percent.

2) There is additional logic involving ROFF_F etc. that is invoked when
marginal-sea balancing is turned on (lms_balance=true). At this time,
this logic is /not/ updated, and hence the lroff_ref_to_local_salinity
option is inconsistent with the lms_balance option; it forces POP to
exit. Note that by default lms_balance=false for the G-cases, and true
for B-cases. So once we want to start running tx0.3v2 in fully-coupled
context, this issue needs to be addressed.

An error message is generated when lms_balance and
lroff_ref_to_local_salinity are both true.