Slight adjustment to the icebergs code
The original icebergs implementation read in an additional record from
the SSS file, whether it wanted to use icebergs or not. This introduced
issues for the tx0.3v2, in which such a third record was not present.
This modification fixes that, as it only searches for the 3rd record if
icebergs are requested. It now also output the sfwf namelist.