Remove locally-referenced virtual salt flux for MELT_F
This commit partially undoes "Locally-referenced virtual salt flux"
commit 7e3afd17564d36f6a406ea30757641661b047358. In particular, it
reverts the locally-referenced virtual salt flux for MELT_F, leaving it
in place for ROFF_F and IOFF_F only. The justification is that MELT_F is
one component of the complex exchanges between ocean and sea ice, and
applying this fix for just on these components (for melting, but not for
freezing) introduces inconsistencies that were not realized at time of
original implementation. This only affects tx0.3v2 configurations.